pyproject.toml
[project]
name = "cyber-ai-block40-lab"
version = "0.1.0"
description = "Small project to introduce students to using AI tools"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai>=2.9.0",
    "python-dotenv>=1.2.1",
]
